backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
aiosqlite==0.20.0
python-multipart==0.0.19
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
pillow==11.1.0
pydantic[email]==2.10.4
pydantic-settings==2.7.1
python-dotenv==1.0.1
aiofiles==24.1.0
websockets==14.1
